requirements.txt
transformers
torch
streamlit
newspaper3k
lxml_html_clean